:sparkles: consumer 및 retry 로직 중복되는 코드 공통화 (#352)
