+++ scheduler/src/main/kotlin/backend/team/ahachul_backend/stream/LostPostConsumer.kt
@@ -1,15 +1,7 @@
 package backend.team.ahachul_backend.stream
 
-import backend.team.ahachul_backend.api.lost.application.port.out.LostPostWriter
-import backend.team.ahachul_backend.api.lost.domain.entity.CategoryEntity
-import backend.team.ahachul_backend.api.lost.domain.entity.LostPostEntity
-import backend.team.ahachul_backend.api.lost.domain.model.Lost112Data
 import backend.team.ahachul_backend.common.client.RedisClient
-import backend.team.ahachul_backend.common.domain.entity.SubwayLineEntity
 import backend.team.ahachul_backend.common.logging.Logger
-import backend.team.ahachul_backend.common.storage.CategoryStorage
-import backend.team.ahachul_backend.common.storage.SubwayLineStorage
-import com.fasterxml.jackson.databind.ObjectMapper
 import org.springframework.beans.factory.InitializingBean
 import org.springframework.data.redis.connection.stream.Consumer
 import org.springframework.data.redis.connection.stream.MapRecord
@@ -25,10 +17,7 @@ import java.time.Duration
 @Component
 class LostPostConsumer(
     private val redisClient: RedisClient,
-    private val objectMapper: ObjectMapper,
-    private val lostPostWriter: LostPostWriter,
-    private val subwayLineStorage: SubwayLineStorage,
-    private val categoryStorage: CategoryStorage
+    private val lostPostUtil: Lost112Service
 ): StreamListener<String, MapRecord<String, String, String>>, InitializingBean {  // stream key, stream value type
 
     private val logger: Logger = Logger(javaClass)
@@ -45,28 +34,13 @@ class LostPostConsumer(
 
         runCatching {
             message?.value?.let {
-                val jsonStr = objectMapper.writeValueAsString(it)
-                val lost112Data = objectMapper.readValue(jsonStr, Lost112Data::class.java)
-                val subwayLine = getSubwayLineEntity(lost112Data.receiptPlace)
-                val category = getCategory(lost112Data.categoryName)
-                val lostPost = LostPostEntity.ofLost112(lost112Data, subwayLine, category, lost112Data.imageUrl)
-                lostPostWriter.save(lostPost)
+                lostPostUtil.convertAndSaveLostPost(it)
             }
         }.onSuccess {
             redisClient.ackStream(streamKey, consumerGroupName, recordId)
         }.onFailure {
             logger.error("Process failed. Keep pending. id=${message?.id?.value}", it)
         }
-    }
-
-    private fun getSubwayLineEntity(receivedPlace: String): SubwayLineEntity? {
-        val subwayLineName = subwayLineStorage.extractSubWayLine(receivedPlace)
-        return subwayLineStorage.getSubwayLineEntityByName(subwayLineName)
-    }
-
-    private fun getCategory(categoryName: String): CategoryEntity? {
-        val primaryCategoryName = categoryStorage.extractPrimaryCategory(categoryName)
-        return categoryStorage.getCategoryByName(primaryCategoryName)
     }
 
     @Throws(Exception::class)
